Add HTTP timing metric
Creates a prometheus Histogram metric to collect timing information on
the request and also the method/status code returned to the user.

This sets up some defaults rules based on the histogram bucket.
Presently this isn't configurable. The metric is added to the default
register for prom-client and will be exported with the default metrics
middleware.